Add central API module
